impl Display for NetworkSize
